+++ components/analytics/metric/api/src/main/java/com/flipperdevices/metric/api/MetricApi.kt
@@ -4,6 +4,6 @@ import com.flipperdevices.metric.api.events.ComplexEvent
 import com.flipperdevices.metric.api.events.SimpleEvent
 
 interface MetricApi {
-    fun reportSimpleEvent(simpleEvent: SimpleEvent)
+    fun reportSimpleEvent(simpleEvent: SimpleEvent, arg: String? = null)
     fun reportComplexEvent(complexEvent: ComplexEvent)
 }